Add links and buttons to the geographic interface
The Dash front page and the Browse All page both now have buttons
which are linked to the Browse Locations page. The records' pages
also have buttons if they have geographic metadata - eventually
these will link to a map with record-specific geoLocations. For
now the button links to the Browse Locations page too.