Add record-specific locations interface and fix nested rectangles
The button on the records page now points to a URL which includes
the docId as well as "browse-locations=yes". This interface is a
cross between the records page and the standard browse page. Only
the locations associated with the record at docId are shown on the
map.

JS now checks to make sure that rectangles do not hide other
rectangles. If rectangle A is found to contain rectangle B, then
A is moved to a layer below B.