Style OC Data Portal results
The OC Data Portal is structured similarly to Browse Locations and
Browse All, but has a different color scheme. The Browse-* pages
are now styled upon matching the browse-page and browse-container
classes, instead of their ID attributes. The ID attributes are
retained in case other JS or CSS rules apply.